Reset this type for covariant subtype proposition on class<T>
Summary:
Some of our entrypoints to subtyping set the `this` type to be the subtype in the environment and leave it up to the inner calls for the premises to unset it e.g. see `simplify_subtype_variance_*` helpers.

The `class<T> <: U` rules were holding onto a stale this type, which made it so, in my example,

https://www.internalfb.com/code/fbsource/[7ec31737b3b7e0f32566dee9f937876b69e692cd]/fbcode/hphp/hack/test/typecheck/class_pointers/deprecated/repro.php?lines=3-24

when we simplify the union for the `$cn` variable under sound dynamic, `(class<C> | classname<I<int>>)`,
https://www.internalfb.com/code/fbsource/[fa420af20785f2c217d0e48bec000edbb32a1bb2]/fbcode/hphp/hack/src/typing/typing.ml?lines=8889-8890
the `C <: I<int>` prop was expanding to `I<tany> <: I<int>` (via the `implements I<this::TInner::T>`). The resulting prop was `(_ <: int && int <: _)`, making is_sub_type_for_union false and leading to an unsimplified union.

Also resets the like-pushing flags since we don't expect to need to push supportdyn or likes through these constructors; rather, the inner type would be constrained.

Differential Revision: D72030464

fbshipit-source-id: d994b02e49015ee6e7b9b3ff1c760445554fddd2