Support `move_TYPE` for Protocol.Value
Summary:
The boost-like APIs for `Protocol.Value` provide const & mutable access to the underlying type, but obfuscate the underlying representation (intentionally).

The API currently allows moving the underlying `T`, but if the type is wrapped, e.g. boxed in a `unique_ptr`, it does not allow moving the wrapper. The `move_<TYPE>()` API allows exactly that.

Reviewed By: thedavekwon

Differential Revision: D71920762

fbshipit-source-id: decafc20ee2ec296573e206d76d9a476f3ab4f15