Make ResponseWriteTimeout don't crash
Summary:
There appears to be a race in the `ResponseWriteTimeout` server test leading to a segfault which, for some reason, is only reproducible on macOS.

The problem is that `SlowReadCallback` gets destroyed before the last call to its `readDataAvailable` method, leading to the crash.

There is already logic to uninstall the callback but it is not used. Use it to uninstall `SlowReadCallback` before it gets destroyed.

Reviewed By: praihan

Differential Revision: D72069227

fbshipit-source-id: 264f870635048259b8dbf97e6a02bf84cfe20009