Fix non-determinism when building FuncFamilies in HHBBC
Summary:
There was a bit of non-determinism leaking into HHBBC when
building FuncFamilies. When we're building a topological sort of
classes in the BuildSubclassListJob, we iterate over a hphp_fast_map
at one point. Since the iteration of a hphp_fast_map is not in any
fixed order, this means the resultant topological sort can be
different between runs. Sort the hphp_fast_map key/values into a
std::vector, and then iterate over that, providing a fixed ordering.

While investigating this, it also revealed a small bug in the
FuncFamily building logic. Depending on the ordering of classes, we
might drop some of the non-regular methods from the method set. This
can be fixed by relaxing one of the checks.

Reviewed By: mdko

Differential Revision: D72186125

fbshipit-source-id: 4f821124f6fa466f895f16625c5cc2545e6d84c4