Fix non-determinism when concatenating together HSL systemlib
Summary:
The script which combines together all of the HSL systemlib together
was processing files in the order which the find command
returns. However, the order which find returns is not fixed (it
depends on filesystem layout). This means the combined file will have
classes in different orders depending on the run. This is mainly
harmless, but it means source locations will be different, which
causes a ripple effect of the compiled repo being different. Fix this
by sorting the output of find.

Reviewed By: mdko

Differential Revision: D72186751

fbshipit-source-id: f789040a44aa63635b5382487c04ca4acdeb20b8