Mark OperationCancelled final
Summary:
This is a prerequisite for any sort of C++26 P2300 compatibility with the "stopped" state, which certainly doesn't support adding context via inheritance.

As a (very minor) bonus, if we know that `OperationCancelled` is `final`, **and** the binary is fully statically linked (no multiple DSOs), then it's possible to have a 3ns RTTI-free test for whether an `exception_ptr` represents an `OperationCancelled`. I won't presently be doing anything about this, but having the option value might prove useful if/when there's bulk migration to value/error/stopped asynchrony.

Reviewed By: ispeters

Differential Revision: D72074521

fbshipit-source-id: f0308bc065a7f04b4d8a4450f92508a515b81cc5