Pass "coercion_direction" when creating unions from typing_subtype
Summary:
In typing_subtype, we sometimes create unions. When we create unions, one of the basic checks that happens during the process is asking if one part of the union is already a subtype of the other (and so then just take the supertype). However, when we check this, we always use the default "coercion_direction," but when we are creating unions from subtyping, we can do more simplifications if we pass the current "coercion_direction."

This diff exposes passes the coercion direction from subtyping into `Typing_union.union` where it can be passed to `Typing_utils.is_sub_type_for_union` and back to subtyping.

Implementation notes:
* I made ~coerce parameters required inside `Typing_union` to make sure I didn't miss anything but left the external API optional.
* I made sure to always pass ~coerce to `Typing_utils.is_sub_type_for_union` inside `Typing_union` but, again, left the paramter optional for other users.

 ---

Motivation:

dlreeves and I have been looking at refactoring the use of type partitions to go from:

```
if ($x is P) { // $x: T; partition(T, P) = (Tleft, Tspan, Tright)
  // $x:#true, Tleft & P <: #true, Tspan & P <: #true
} else {
  // $x:#false, Tright <: #false, Tspan & !P <: #false
}
```

to

```
if ($x is P) { // $x: T; partition(T, P) = (Tleft, Tspan, Tright)
  // $x:#true & P, Tleft <: #true, Tspan <: #true
} else {
  // $x:#false & !P, Tright <: #false, Tspan <: #false
}
```

(This is D68849503)

However, the type solver doesn't behave very well with intersections of tyvars. In order to address this, we expand the types that we recognize as exact negations for the transform: `A & B <: C => A <: C | !B`

Expansion of this transform itself results in running into additional incompleteness, especially for `#0 & !P <D: dynamic` which when transformed into `#0 <D: P | dynamic` isn't always handled well. (See a current example of this in the new test "hphp/hack/test/typecheck/dynamic_or_null_is_dynamic_under_dyn_assumptions.php"

However, when `P <D: dynamic` we can transform `#0 & !P <D: dynamic` into `#0 <D: dynamic` instead.

 ---

This diff introduces errors in a few files in WWW, but they seem to be the result of now directing the code into other incompletenesses, e.g. when minifying them, I found false-errors for similar code that occur when running against the prod-typechecker: T215988118, T215988614

So this is a trade-off on incompletenesses.

Reviewed By: dlreeves

Differential Revision: D70726155

fbshipit-source-id: 1d5f00b7c1254892468513d2b2bcaf67f52d752b