fix iteration of mangled-name field
Summary:
In python, `self.__{{some_attribute}}` is resolved to `self._MyClass__{{some_attribute}}`. This is called class mangling. Whatever dark magicks power cython manage to bypass it, but for normal Python we have to accomodate it.

Here's an approach to make this work with in-place migrate. It does't add any overhead to the non-exception path.

Reviewed By: yoney

Differential Revision: D72010620

fbshipit-source-id: 0441582503b630ce4a38cfeaaf0b121970b18581