pass down location hint when creating custom compressor
Summary:
`CustomCompressorFactory` is a challenging library to mock and test, because, when running the integration test locally, both the client and the server uses the same instance. This makes it not ergonomic to inject failures only on clients or only on servers.

In this diff, we pass down some additional location hints. These should have minimal performance overhead, but would allow us to inject failures into the right places in integration tests.

Reviewed By: cevans87

Differential Revision: D70809362

fbshipit-source-id: 98c23408ebda201425d45d202bcef46f4dc9c751