address property field name collision
Summary:
The in-place implementation uses `property` for field / attribute access. It blow up when there's a thrift struct with the field name `property` because subsequent usage of `property` starts to refer to class-local `property` not the python builtin.

The fix is easy enough: mangle the builtin and use that in the tempates.

Reviewed By: yoney

Differential Revision: D71875833

fbshipit-source-id: 5952d11d5e84053425a00c2d698d632df55178bf