update the repo build to support packages, force symbolrefs and --exclude-xxx options
Summary:
This diff updates how the build is performed, according to the options below.  If Packages V2 are not enabled this diff is a no-op: the parse, index and emit phases are not changed.

If Packages V2 are enabled with `--vRuntime.Eval.PackageV2=true`, the build is influenced by the following options:

- `--vRuntime.Eval.ActiveDeployment="xxx"`:
  - if no ActiveDeployment is set, the build uses SymbolRefs;
  - if the ActiveDeployment is set, the build uses the Deployment specification for `xxx` in the `PACKAGES.toml` file.

- `--vRuntime.ForceEnableSymbolRefs=true/false`:  if an ActiveDeployment is set, `ForceEnableSymbolRefs` causes the build to include all the ActiveDeployment files _and_ all the SymbolRefs files.  If not, only the files specified by the ActiveDeployment are included in the build.

Additionally, the build is controlled by these options:
- `--exclude-dir`: files excluded with `--exclude-dir` are excluded from the build _unless_ SymbolRefs or Package rules pull them in.
- `--exclude-pattern` and `--exclude-file` (unused by `hphp_files.mk`): files excluded with `--exclude-pattern` and `--exclude-file` are _always_ excluded by the build, even if SymbolRefs or Package rules pull them in.

Remark: the option `--report-ondemand-edges` only lists files included by SymbolRefs.  Logging files pulled in the build by Packages will be handled in a separate diff.

Reviewed By: paulbiss

Differential Revision: D66827456

fbshipit-source-id: dafb547b7a5d237420deea495cdd3ee4c83cecd4