Don't forget to check arguments
Summary:
This diff resolves a typehole related to object creation.

We have logic to resolve the constructor for `new` expressions. The arguments are typechecked with respect to each possible resolved class. When no such classes are found, we default to typechecking the arguments independently to discover errors such as "unbound name" errors.

This worked well until we did true types which led to `dynamic` being added to the resolved type list. As a result, the empty list case wasn't hit and subsequently arguments weren't checked independently.

This diff fixes this by growing the default case to just having a `dynamic` type as a result of resolution.

Reviewed By: viratyosin

Differential Revision: D72281656

fbshipit-source-id: 156d49a8f9e595b30e59d1cd81893d0bf5170b3a