Expose registered commands from folly::NestedCommandLineApp
Summary: To test a NestedCommandLineApp, it's useful to be able to see the registered commands, and generate a variables_map that contains all the relevant defaults for a particular command. At the moment, the set of registered commands (an their associated options_descriptions) are not accessible. Adding a simple getter here to expose them.

Differential Revision: D72258645

fbshipit-source-id: 4766eb1f71fb505d51cf1fc0f8f28b8db5a49c95