write files in a package build to a file
Summary: Add the `--report-files-in-build=<filename>`.  If set, the names of all the files pulled in the repo at repo-build time are listed in <filename>.  The list is exhaustive, contrarily to the list of files in `--report-ondemand-edges`: the latter is generated from dependencies between symbols, and misses files that have no dependencies.

Reviewed By: WizKid

Differential Revision: D67544455

fbshipit-source-id: f6bdee5ed5aff581c12cddd0f8ac55b47ae9a79a