Use a monotonic clock in ThreadFactoryTests
Summary: Use `steady_clock` instead of `Util::currentTime` which gives a platform-specific choice of `clock_gettime(CLOCK_REALTIME, ...)` or `gettimeofday`. The former is monotonic and, unlike the other two, suitable for reliable measurement of time intervals. Neither of the current choices are monotonic, quoting the docs "The time returned by `gettimeofday()` is affected by discontinuous jumps in the system time".

Reviewed By: iahs

Differential Revision: D72351371

fbshipit-source-id: 9e6e21259f59e90a139ad31e964cb2fc4d2d42d0