IoUringProvidedBufferRing: madvise MADV_NOHUGEPAGE to workaround kernel v6.9 pin user pages bug
Summary:
There is a known issue in v6.9 when pinning huge user pages that causes it to EFAULT. The bug is in an optimisation for handling large folios. The error is propagated back to userspace and throws an exception in IoUringProvidedBufferRing.

The bug is fixed in v6.11 and above. But to workaround the issue in v6.9, madvise MADV_NOHUGEPAGE and disable huge pages to avoid the buggy optimisation.

Reviewed By: yfeldblum

Differential Revision: D72265355

fbshipit-source-id: f3a359e1e470bf1344f41bea9418ed6a42a49a42