Fix coeffect case in D69159959
Summary:
We want to allow un-annotated `__Memoize` to be allowed if annotating the `__Memoize` function would be disallowed due to coeffects.

For example:

```
<<__Memoize>>
function memo_fn($a, $b)[leak_safe]: string {
  return "args: $a, $b \n";
}
```

If we changed it to `<<__Memoize>>` to `<<__Memoize(#KeyedByIC)>>` (or any other flavor of memo), it would raise another hack error. This creates code that has no validate annotation configuration.

Reviewed By: jano

Differential Revision: D72336343

fbshipit-source-id: c3a59d9ece5ead360aefb87367f411ad95337637