pdu: add PartialEq to Clock
Summary:
Reading https://www.internalfb.com/intern/staticdocs/watchman/docs/clockspec, it
is my understanding that clients should merely use the clock as returned by
Watchman, and for the most part their internals should be opaque. However, it
feels reasonable to offer library users to be able to tell if 2 clocks are
identical or different.

Reviewed By: genevievehelsel

Differential Revision: D72464932

fbshipit-source-id: 7d3e0eb80c095f91ae90ee55bc46e6001347d7c1