Weak.h a solution to everyone linking libpython for code inspection in c++
Summary:
We have this issue there are some core C++ libraries that dep on libpython, causing C++ binaries to have dependencies on python when the service owners are unaware why their C++ service now needs libpython3.12.so when to their knowledge they don't use python at all.

The goal of this project is to provide weak symbols and a unified way to tell if python is linked or not.  That way the core c++ libraries need not themselves link libpython, but if it is linked they can use it in limited ways.

see: S504477 and https://fb.workplace.com/groups/pyupgrades/posts/3601313680172758/ for justification

Reviewed By: parulgupta1004

Differential Revision: D72285611

fbshipit-source-id: 4d0820f2d37e1c65eb6e415d1c7aa01d901e1f8e