Fix ClientOnlyTimeouts test
Summary:
The `ClientOnlyTimeouts` test is extremely flaky.

It almost always fails on macOS:

 {F1976800949}

It fails on Linux with tsan in ~80% of cases:

 {F1976800993}

It occasionally fails on Linux without tsan:

 {F1976801019}

The reason why this test is so flaky is that it consists of four test cases that interfere with each other via a shared server thread. So the sleep from an earlier test case that checks the timeout may continue blocking the test case after it, causing that test case to also timeout:

```
fbcode/thrift/lib/cpp2/test/server/ThriftServerTest.cpp:3026: Failure
Value of: shouldTimeOut
  Actual: false
Expected: true
```

This becomes even more obvious if we increase the sleep from 50ms to e.g. 100ms. If there was no interference we would expect the test to be less flaky (or at least stay the same) because of the increased margin for failure. However, the test becomes more flaky, now failing almost all the time on Linux.

Fix it by not reusing the server thread between the test cases.

Reviewed By: praihan

Differential Revision: D72461871

fbshipit-source-id: 3934fdc152a64e30069810b3de7b6ff14bb066bb