Add ValueHolder as a type-erased Value wrapper
Summary:
## Object  <-> Value

`Protocol::Object` & `Protocol::Value` as thrift types contain a birectional dependency. A thrift `Object` contains fields of arbitrary `Value`, while each `Value` itself can also represent an `Object`.

In the C++ world, this mean we have to type erase one of the sides. In the `protocol_detail.thrift` definition an `Object` variant is boxed within the `Value` union. (Note: the map storing the fields for an object is also boxed, not for type erasure, but for reduced stack size (~perf)).

`experimental::Object` instead chooses to erase the type while keeping it on the stack by introducing the `ValueHolder` class. This type conforms to the static layout of a `Value` & can trivially be convertible to it.

## API

Both `ValueHolder` and `Value` implement the `ValueAccess` API & can therefore access the value union.

Reviewed By: thedavekwon

Differential Revision: D72348295

fbshipit-source-id: 1948b40e3581ee00f3468c7a3c29e6ad9e719ad7