Flip the default
Summary:
Flip the default on strict switches to enable it by default. This will allow us to correctly exercise every test.

This diff contains updates to test cases, here's the breakdown:
* Everything under `hphp/hack/test/typecheck/` was known and covered by the `typecheck:verify_typecheck_new_exhaustivity`.
* `hphp/hack/src/hackc/test/infer:typecheck_instr` was failing so added a `default` case in an Infer test to prevent errors.
* `hphp/hack/test/refactor_sd/dump_constraints:dump-constraints` failed to generate constraints because the switch test cases needed defaults but didn't have any.
* `hphp/hack/test/sound_dynamic/typing:good` had two test cases with missing defaults on `int`s, so we changed the test case to make them well-typed.
* `hphp/hack/test/tast:tast` one TAST expected file is updated to have legitimate switch errors
* `hphp/test:typecheck_run_php` runs the typechecker over the HHVM test runner, this had non-exhaustive switches, so added FIXMEs.

Most importantly, this diff is right on top of D72329724 which fixes the fanout bug for strict switches. Prior to being on top of the fix, flipping the default caused two integration tests that caught the bug to fail:
* `hphp/hack/test/fanout:fanout`
* `hphp/hack/test/incremental:enum_exhaustiveness`

Notice that this diff makes no changes to these tests, but now they are passing confirming that the diff below fixes the fanout bug.

Reviewed By: vassilmladenov

Differential Revision: D72342567

fbshipit-source-id: 4a5ca8d8e9e6af2fe03b83d20ab14820fc826043