Allow __SimpliHack attribute to accept nonliteral expressions
Summary:
# Motivation
In general the arguments to attributes in Hack are limited to literal expressions. This is because there is only a limited set of expressions that can be represented by the runtime at that position. This restriction leads to unnecessarily complex work arounds for invoking promptlets for SimpliHack. Today to invoke one requires writing `__SimpliHack(PromptletClass::class, arg1, arg2)`. After this change it is possible to support `__SimpliHack(promptlet(arg1, arg2))` instead which looks nicer.

# Changes
In this diff we special case the __SimpliHack attribute to allow nonliteral expressions to appear as arguments. For the runtime we make this work by stripping out the attribute completely. In the type checker we skip validation of arguments for SimpliHack attributes.

I'll make the change to the SimpliHack interpreter to support the nicer syntax in a subsequent diff.

Reviewed By: patriciamckenzie

Differential Revision: D72272575

fbshipit-source-id: 07d21d2a6a547e6eb8697e3ec3fff67212b59974