Support top-level function in interpreter
Summary:
This adds support for top-level functions. This combined with the other diffs in the stack allow for a nicer syntax for calling promptlets so it resembles top level function calls, i.e.

```
<<__SimpliHack(DeriveBuilderFor(Person::class))>>
```

# Implementation
I evaluate `Id()` expressions to `FunctionPtr` values which wraps `Nast.fun_def`. I then defined a converter to a `Value.callable` which contains the common fields for `Nast.fun_` and `Nast.method_` that is used by `Call.eval`

Reviewed By: patriciamckenzie

Differential Revision: D72283936

fbshipit-source-id: 39fe53a6cf4468c94c59241318a645ea6771fef5