Fix request_channel_test_resource_pools failure on Mac;
Summary:
This diff fixes a test failure on Mac for the `RequestChannelTest` class.

Due to the difference in Linux and MacOS, when the server is missing, the exception message on Mac is `apache::thrift::transport::TTransportException: AsyncSocketException: connect failed, type = Socket not open, errno = 60 (Operation timed out): Operation timed out`, and it is `apache::thrift::transport::TTransportException: transport is closed in write()` on Linux.

So this diff changes the error string from `transport is closed in write()` to `TTransportException` to make it work on both Mac and Linux.

Reviewed By: robertroeser

Differential Revision: D72669982

fbshipit-source-id: e17218728e6e9a1563877fa319a88d829479682a