produce true and false assumptions for case types
Summary:
Case types may have optional where clauses.

During refinement, we want to be able to take advantage of the fact that if a variant is conditional on some where clauses, then in a branch where a value is known to be part of that variant type, those clauses must be satisfied.

This diff adds the where clauses as assumptions to a variant's sub-partition when splitting case types. The way the assumptions of sub-partitions are combined, it falls out that assumptions of variants that do not match the predicate are ignored and when multiple variants may match a predicate no assumptions are taken (since we ignore non-trivial disjunctions).

This diff also extends the assumptions tracked to include false-side assumptions in addition to true-side assumptions.

For case types with where clauses, this is needed to so that you can conclude the where clauses associated with the final variant when testing variants exhaustively without the need for an explicit and redundant type test. For example:

```
case type CT<T> =
| int where C1
| string where C2
| bool where C3;

function foo<T>(CT<T> $x): void {
  if ($x is int) {
    // assume C1
  } else if ($x is string) {
    // assume C2
  } else {
    // we want to assume C3 here
    // without requiring an explicit `if ($x is bool)`
  }
}
```

Because of how previous refinements' are tracked via intersection of negative predicates, we need to make use of these intersected types in order to filter out which variants' to exclude so that eventually, only one variant remains to provide a false assumption when doing the final partition (for `$x is string` in the example above)

Reviewed By: dlreeves

Differential Revision: D64611956

fbshipit-source-id: edfa11e4f416a841ebc78e6fcd4cd8a5e000aab3