Add unsafe flavor of classname_to_class for sandboxes
Summary:
Adds an undocumented backdoor to `HH\classname_to_class()` to allow resolving a class from a string without it being marked `<<__DynamicallyReferenced>>`. This backdoor is only meant for sandbox mode, and it will halt the build with an `always_assert` in HHBBC should it somehow land to a prod path. I also added assertx's in the JIT and runtime just as a sanity check.

Hack will not expose this extra parameter, it will raise an error about too many parameters.

Reviewed By: jano

Differential Revision: D72769800

fbshipit-source-id: 7fc87b2e56d0caf64554b6208deaafda6a21e941