Forward fix memory allocation change
Summary:
D71488580 accidentally changed the pattern of memory allocation. Instead of allocating a series of small (roughly 16KB) IOBufs, it changed the behavior to allocate a single large buffer for each list of integers. This is problematic if the list size is very large. We don't want to call malloc(100MB).

This diff restores the previous memory allocation pattern, but keeps the performance wins.

I am submitting this as a stand-alone change because, as a forward fix, it is more urgent to land than my other changes.

Context: it turns out that `ensure()` always gives you the requested number of bytes in a contiguous buffer. TODO provide code pointers.

Reviewed By: yfeldblum

Differential Revision: D72823267

fbshipit-source-id: 4d9d775a2f5bec1d379a574e2dd3fc336fcba138