Use netty’s bytebuf based base64 serializer to enable garbage free serialization
Summary: When base64 serializing thrift structs we use ByteBuf’s internal which are garbage free and released after use. The intermediate bytes can create heap waste as they are immediately thrown away. In this diff we use the intermediate ByteBuf and pass it to Netty’s base64 serializer that also creates a temporary ByteBuf for writing the Base64. This can be converted into a byte[] or String depending on what is needed. This should reduce heap waste for those using base64 serialization.

Differential Revision: D71921376

fbshipit-source-id: 3183cd31c4f8088316cc6ce8d3459a065b355416