mangle names to avoid name pollution
Summary:
## What

- Mangle names when converting from notebook -> hack to avoid name pollution for source-controlled notebooks
- Demangle when converting hack -> notebook

## Why

The goal of this stack is to do rewriting so that we store source-controlled notebooks in WWW with rewritten names to avoid name pollution.

Since names are unique in WWW, we don't want someone definining `function add1` in a notebook to make it so that no one else can write an `add1` function (in notebooks or otherwise).

For example, given notebook ID N1234, we do the following rewrites:

`my_fun` ~> `n1234_my_fun`
`gen_foo` ~> `gen_n1234_foo`
`MyClass` ~> `N1234MyClass`

[more context](thttps://fburl.com/source-controlled-notebooks-for-hack-plan)

## Follow-ons

- Consider additional measures of containing things-defined-in-source-controlled-notebooks: such as filtering out names-defined-in-notebooks from completions and filename conventions

Reviewed By: lillyfwang, tsurupin

Differential Revision: D72863598

fbshipit-source-id: b0252dc131d5c0a9b4a48913d35fe53bcdee8f59