Add NativeSet support
Summary:
This diff introduces `NativeSet`, a tagged representation of various `folly::F14FastSet` based specializations of `Value`.
In the `Protocol::Value` implementation, a set was always represented as `folly::F14VectorSet<Value>`, significantly increasing the memory use & reducing cache efficiency for primitive types.

With `NativeSet` we provide specializations for lists of all primitives, e.g. `set<i32>`, `set<string>`, ...
so that consumers opt-in to the native equivalent, at the cost of resolving a single external tag for each set (in contrast to a tag for each element).

Reviewed By: thedavekwon

Differential Revision: D72358633

fbshipit-source-id: 7537f477ce2cb89de74e18fc6b423a02832b0953