Disallow using Ordinal as Id in `patch()`
Summary:
# TL;DR

There is a bug that when Ordinal is used in `patch()`, we might patch incorrect field.

This diff banned Ordinal in `patch()` since it doesn't look like this is actually used in production.

The bug will be fixed in the diff after next diff in stack.

# Detail of the Bug

Currently we generate the following fields for struct patch

```
struct MyStructPatch {
  1: optional MyStruct assign;
  2: bool clear;
  3: MyStructFieldPatch patchPrior;
  5: MyStructEnsure ensure;
  6: MyStructFieldPatch patchAfter;
}
```

The bug is that inside `patch()`, we converted `Id` to ordinal later `op::get_ordinal_v<FieldPatchType, Id>`. (Here FieldPatchType is MyStructFieldPatch).

The Id should be based on `MyStruct`, not `MyStructFieldPatch`.

* If Id is ident or field id, this is fine since fields in `MyStructFieldPatch` and `MyStruct` have the same field name and field id.
* However, if Id is Ordinal, `MyStruct` might have unpatchable field (e.g., `shared_ptr<const T>`) that doesn't exist in `MyStructFieldPatch`, thus Ordinal will be incorrect.

#buildall
#buildonlynotests

Reviewed By: praihan

Differential Revision: D72922720

fbshipit-source-id: d5dd3a90d852cc0da02c858b4fc5b924efdc9eef