Move ensure_folly_init sooner in hh_single_complete
Summary:
## What

ensure folly is initialized sooner

## Why

Attempt to fix an error someone is getting when updating build rules with `hh_single_complete`. D72832800. I don't want us to be opted in to some weird mode of folly for *all* OCaml binaries and libraries just because of something we use only in tests.

## Risks

Not really anything. `hh_single_complete` is only used in tests.

Reviewed By: viratyosin

Differential Revision: D73040633

fbshipit-source-id: 8d92425935effc6ce8ae10bf8715fac5fef38f3e