Better support for the resume function
Summary:
The `resume()` function was originally intended to match 1:1 with `pause()`, but this is not perfect as it would be nice to unconditionally resume in a destructor if the work is paused.  Having a separate variable - or even checking the operation's state - leads to possible race conditions.  Instead, change the function to just always resume if the state is `WaitForConsumer` and don't do anything in any other state.

Also, verify that `impl_` is not null when calling the top level function.

Differential Revision: D72884849

fbshipit-source-id: 3572ec174cd29f899225dc89d1622584be435e58