non_hack cell is stored in comments
Summary:
## What

Support non-hack cells in notebooks.
They should be stored in comments.

The only way we can tell is by parsing the bento metadata for the 'language' field. If the metadata is missing or in an unexpected format, we assume it's Hack.

## Why

Would be nice to have (weak) support for non-hack cells in source-controlled notebooks.

The benefits of source-controlling Hack all come from standard tooling (codemods, type-checking, testing) working with the Hack source code. We're unlikely to teach these tools to understand SQL cells, graphQL cells, etc. any time soon, so just put them in comments when source-controlling.

The desired behavior (non-Hack cells in comments) is in accordance with [the design](https://fburl.com/source-controlled-notebooks-for-hack-plan) but not implemented until now.

Reviewed By: nt591

Differential Revision: D73044959

fbshipit-source-id: a7f69ea384b06a9e98b5445f5aedf4f32390dab4