fix name demangling when converting from hack (.php) to notebook JSON
Summary:
Fix a bug with name mangling, introduced in D72863598 (see that diff for context).

We were correctly demangling class names when converting from Hack (.php) to notebook JSON (.ipynb), but I messed up the logic for function names.

What this means is that if someone were to switch between editing a notebook as .php and editing a notebook as .ipynb, they'd keep accumulating longer and longer prefixes to function names. This is bad.

Reviewed By: nt591

Differential Revision: D73058576

fbshipit-source-id: a92663a8a288ee9418a8c20b5d4843546ca14bac