Track Dependencies During Evaluation
Summary:
sim
We want to add dependency awareness to SimpliHack. By dependency awareness we mean that if the context used to generate code changes, SimpliHack should make the user aware and ask them to re-generate their code.

To enable this we need to record dependencies in our dependency table so we know what inputs went into computing the prompt. This is what this diff does. We make the following changes.

1. We change the API for `Simplihack_interpreter.eval` to take a `Tast_env.t` instead of only the `Provider_context.t`. The `Provider_context.t` is contained within the `Tast_env.t`, but it also includes the `Decl_env` that contains the dependent root. When evaluating a Simplihack promptlet we add dependencies to this root, so that if any of these dependencies change we will retype check the file contain that dependent root.

2. We updated `Simplihack_prompt`'s visitor to pass in the `Tast_env.t`. I discovered while implementing this we weren't setting the dependent root when visiting module definitions. I fixed that in this diff as well

3. I added a dummy tast_check in `simplihack_record_dependencies.ml`. This is added to run the interpreter during type checking to record the dependencies. This is put in temporarily so we can test that the dependencies are being added properly. In a future diff we will exchange this for a real tast_check that can report errors.

Most notably we introduced a new variant in `Typing_deps.Dep` called `File`. This dependency indicates that a dependent depends on the contents of another file. We need this because most other dependencies only take dependencies on the declarations that appear in the file. However SimpliHack will look at function bodies during evaluation. This means if the contents of the body changes we want to retype check the file that invokes the promptlet.

Differential Revision: D72679042

fbshipit-source-id: a1a51faf07d9005cec4802f4bbd0e993f315432e