Update System Prompt to Update Hashes
Summary: When the SimpliHack code action is run, we want it to update the hash associated with it. To do so we expand the system prompt to include the hash of the prompt as well as examples of how to apply it. A suitable AI agent should be capable of following the instructions and updating the hash for the user.

Differential Revision: D73077726

fbshipit-source-id: 8f021b3f3e967c9e88febce191d7b478acd3ea96