Report errors to run prompt
Summary:
This diff builds on the other diffs in this stack by introducing a new tast check for SimpliHack. This replaces the temporary simplihack_record_dependencies.ml by adding a proper tast check. For now this check visits every user attribute looking for instance of `__SimpliHack`. If found we check if there is a single argument. This means SimpliHack has not ran yet, so we generate an error instructing the user to run SimpliHack.

If there are two arguments we assume the second argument is the hash of the prompt. We then evaluate the first argument and if it evaluates to a string, hash it and compare to the hash stored. If there is a mismatch we report an error and suggest running SimpliHack again to regenerate the code.

Note that there are other checks we could make, but haven't decided yet if they should be done as a tast check or part of typing. For now we only do this check to demonstrate dependency awareness.

Differential Revision: D73076289

fbshipit-source-id: 1808950117a82587977701362951c8cc35011878