Implement Fanout for File Dependencies
Summary:
This diff integrates the newly added file dependency in D72679042 to our fanout procedure. The intention is if a file changes at all, we trigger rechecks for all dependents of that file.

This is conceptually simple, but is tricky in practice since the fanout logic is not well abstracted. In particular the place where the initial set of files that changed is reported and when we compute fanouts depending on changes to folded declarations are far apart. I believe the best place to do it is within the fanout model and passing in the list of reparsed files. It was difficult to trace through the code to determine if every changed file will show up in reparsed, but seems good enough for now.

Differential Revision: D72829481

fbshipit-source-id: dc880c5fce11d8ffe3afc64cb952590a6bf8cd7e