Add method to update stack offsets for all relevant IRExtraData (HHIR stitching part 0)
Summary:
I need a way to update the stack offsets during cloning of instructions for the inlining pass. That's because the caller's stack depth might be deeper which need to be adjusted so instructions don't clobber unrelated stack values.

The update method checks for if the `IRExtraData` has a method `updateStackOffsets` and if so invokes it, otherwise it's a no-op. I also needed to add a non-const version of `instruction->rawExtra()` so that it can be modified.

In the `updateStackOffsets`

- If an offset is a `IRSPRelOffset` we add the delta (a negative number) to make depth larger
- If an offset is a `SBInvOffset` we subtract the delta

Reviewed By: paulbiss

Differential Revision: D72866077

fbshipit-source-id: fa78775df97f137720885a9e7735a2379f21370c