Add AF_VSOCK support to SocketAddress
Summary:
This makes it possible to construct AF_VSOCK addresses.

This is implemented by adding another type into the variant, see previous diff for the introduction of that.

Note that the port is widend to 32 bits, as VSOCK ports are 32 bits. We use new methods to return the full width port to avoid accidental narrowing.

Reviewed By: yfeldblum

Differential Revision: D72569916

fbshipit-source-id: 73dcddcde4839b1b6f70e87923e687a5e25ced21