float field always returns `float`
Summary:
The thrift-python `TypeInfo` currently accepts any instance of `int` on python `__init__` or `__call__` for setting a thrift `float` field. On accessing the set field, the thrift-python struct returns the *original* Python object, which is not necessarily a `float`. This violates the thrift spec, especially because the object's state changes across a roundtrip serialization, as documented in tests.

This diff continues to accept instances of `float` subclass when creating the struct from Python, but it will immediately convert it to `float` for thrift-spec compliance.

Reviewed By: createdbysk

Differential Revision: D73014964

fbshipit-source-id: 73b54810c517c0b0b938803d56875b1a4e1e7041