Remove read_first/read_half in NativeObject bench
Summary:
Remove the `SparseAccess` support for the `Protocol.Object/Value` benchmark.

This behavior attempted to measure/profile the cost of traversing a thrift hierarchy while only performing partial access (either the first property, i.e. `read_first`, or half of the properties, i.e. `read_half`, for each level.

Unfortunately the cost heavily depends on the content & layout of the data, in addition to being a non-trivial comparison to the `NativeValue` implementation which has a higher baseline cost of access at each level due to the `std::variant`.

Reviewed By: thedavekwon

Differential Revision: D73038542

fbshipit-source-id: 923577ebc9b3e34ecce7d63f8935fac844fb0aec