Add property types tests
Summary:
# Problem
`test_property_types_hint` in the `abstract_types_test.py` checks whether the generated property type hint for properties in thrift abstract types are correct for the thrift field types that they correspond to. While this appears to be correct, this does not correctly handle cases where the type that pyre infers doesn't match the property type hint. This happened when there was a time that pyre returned `typing.Any` for any property.

The original reason we chose this approach was because at that time there wasn't any way in CI to confirm that the assignment to an incorrect type, which requires a pyre-ignore when that type check works correctly, renders the pyre-ignore unused if the assignment to the incorrect type doesn't cause a type mismatch. Now there is a CI check for unused ignores.

# The correct approach
Tests must test usages that will appear in code. The correct test for the property types is to assign the properties to type-hinted variable and confirm that the assignment matches or does not match types.

With the CI check for unused ignores, if the pyre-ignore for the type mismatch case becomes unused, then the CI check will catch it.

Reviewed By: ahilger

Differential Revision: D73128388

fbshipit-source-id: 298700af9d782da38ec44b82371650e93a27cb1f