Move whisker/tree_printer.h to thrift/shared
Summary:
This diff introduces a new `thrift/shared` directory that should contain code that is intended to be shared across `thrift/compiler` and `thrift/lib/` (i.e. the runtime).

With DynamicThrift / RuntimeSchema, we expect that much of the compiler APIs will end up being available at runtime. Therefore, this shared directory is expected to grow over time.

For now, we start off with (a rather inconsequential) addition of `tree_printer.h`. My intention is to re-use the `tree_printer` implementation for debugging `SyntaxGraph`.

Reviewed By: sadroeck

Differential Revision: D73162356

fbshipit-source-id: ce533df30553cc13655bacc357d27487b88faab2