Create test target for SimpliHack interpreter
Summary:
Currently the only way to test the output of the interpreter is via a ide_code_action test. This is less than ideal because it includes the system prompt and requires using a comment to set the cursor.

Instead I've added a new test target that tests the output of the interpreter directly. This is done by adding another flag that can be passed to hh_single_type_check. Right now just added a simple test, but plan on adding more as I harden the interpreter.

Differential Revision: D73204339

fbshipit-source-id: b8945ae6764e9afcd34cd5c690b199c168c849c1