Add Ability to Read File Contents
Summary:
Promptlets will be made of largely text. If the text gets large it is desirable to write this in a separate file. We add the ability to do this within SimpliHack by introducing specially interpreted functions, in this case `\HH\SimpliHack\file()`

This function takes a string which is a relative path from the hh_server root, and reads the contents of the file as a string. This function should only be invoked within a promptlet, so I've also introduced a new context `HH\SimpliHack` that is required to invoke the function.

In order to test this I changed the syntax that multifile supports to also allow specifying `~root/file`. The `~root/` prefix denotes that the file path should be interpreted as being a relative path from the root. Otherwise the prefix would default to the `Relative_path.Dummy`

Differential Revision: D73277164

fbshipit-source-id: 7c656e3bcaff80c3c2fadd1dde7099feb23bc274