Make type const reified check not check bounds
Summary: This check was incorrectly restricting some type constant definitions as not being reifiable on the basis of types in their bounds, when we should only care about the value; example in the prior diff. It was also walking them from the caller perspective, when all it should need is the ttc_reified info.

Reviewed By: dlreeves

Differential Revision: D73154439

fbshipit-source-id: bee32400453ad09d2bb792d3a9fd228e0993f3ab