Add IPureMemoizeParam, a pure version of IMemoizeParam
Summary:
getInstanceKey() predates purity, but was always supposed to be pure.
Unfortunately it is not, which makes serialize_memoize_param() also impure.
This is a problem, since serialize_memoize_param() is used implicitly by HHVM's
<<__Memoize>> and implicit context features. Lack of coeffect checks creates
a backdoor.

An attempt to lift serialize_memoize_param() from implicit context's C++ logic
to Hack logic reveals this problem, so let's do the first step of addressing it
by providing a pure interface that the relevant logic can migrate to.

Reviewed By: WizKid

Differential Revision: D73226927

fbshipit-source-id: 4586702fdbe74caba2374244dabf01027bbeae4d