Add support for low priority awaitables to asio scheduler
Summary:
Add asio scheduler support for low priority awaitables. If an awaitable is marked low priority via `Create<true>` it will be enqueued into a low priority context. The scheduler will always process non-low pri awaitables before looking at the low priority context. This can be used to mark some awaitable tree as lower priority to give preference to more latency sensitive work. The semantics of execution are largely unchanged (awaiting a child will still lift it into the parent's context), it's just the order of execution when there are multiple work items to process.

An implementation detail is that the low bit of the context_idx_t is used denote the priority of the wait handle (1 being regular, 0 being low). In practice there are never more than a handful of contexts so this isn't a huge loss, and allows quick lookup of the wrapper index (>>1), or priority comparison (low bit).

Reviewed By: jano

Differential Revision: D69574242

fbshipit-source-id: 0098261f701726f19d4595c9e8584a75d278d85a