Fix bug: find-references showed too-few results when called repeatedly
Summary:
## What

Fix a bug where, from the user's point of view, calling find-refs on the same function would sometimes lead to progressively fewer results.

**Before** (bad)
 https://pxl.cl/742rW

**After** (good)
https://pxl.cl/742ss

## Explanation of bug

Streaming find-refs combines results from Glean and from our own logic for dealing with "open files" in IDE.
The bug was in our logic for calculating references for "open files": which in this context seems to mean "files for which we have state in hh_client".
The very act of retrieving a reference can end up adding the file containing that reference to "open files" which lead to what the behavior the user was seeing.

From the point of view of internals:
- we receive a request for references for C::m
- we retrieve references from Glean, and their corresponding files are added to "open files"
- we receive a second request for references for C::m. This time more files will be in "open files" and so we will use `hh_client` to find references rather than using Glean. The bug in `hh_client` (fixed in this diff) manifested as fewer references being provided than should have been provided.

## The fix

The algorithm for requesting references for a given class+member from a single file was simply wrong.
I think the author of the code probably thought they were using an API that would provide *all* descendants of a class, when instead the API only provided all descendents of a class *defined in a given file*. The core of the change is a one-liner in FindRefsService:

**Old algo**:

```
for each open file `f`:
    let `class_set` = `C` + (all descdendants of `C` defined in `f`)
    for each use of a member in file where the member name matches:
        if the receiver is in `class_set`, use this reference
```

**New algo**:

```
for each open file `f`:
    for each use of a member in file where the member name matches:
        if the receiver inherits from `C`, use this reference
```

## Considerations

There may be cases where the new algorithm is slower. That said, the old algorithm was so very wrong and there's no obvious way to speed up the correct algorithm.

Differential Revision: D73433872

fbshipit-source-id: d07cf40c1be7304e11b12e3b6508556fc63bc3ab