Do not check types of out-only parameters
Summary:
Input values of out-only parameters is ignored and replaced with a default
value for the given type. Stop enforcing types of inputs, since they are not
being used.

This does nothing for most <<__OutOnly>> parameters, since they are typed as
mixed. There are a handful of parameters that are typed better, we will be more
permissive about passing ignored garbage to them after this diff.

Differential Revision: D73407684

fbshipit-source-id: 8809671eed0083232cf052b75d2b8c79cc3299f0