Improve error handling on undeclared struct initialization
Summary:
# Context

This came up in my usecase. See unit test for clarity. When one initializes a const with a struct initializer, but with say a typo in the struct type, there's an ugly crash due to this assert https://fburl.com/code/lmxzm9za .

# This diff

Updates so that instead of a crash, we display this issue in the diagnostics. Also included the expected type in the err msg for extra clarity.

Reviewed By: iahs, vitaut

Differential Revision: D73141930

fbshipit-source-id: 88ce121b123ea52a0c8e782a166e045ea9afbf41