Use xplat rustc for generating Cargo.lock and .rustc_info.json
Summary:
Autocargo is generating different output on different devservers. Since D68684463 + D73042576, the `generate_cargo_lock` step is sensitive to whatever default "rustc" is in your $PATH. For edward-shen in D73070247 that was "rustc 1.86.0". For vmagro in D72982039 it is "rustc 1.85.0-nightly". On my devserver it is "rustc 1.88.0-nightly".

This diff tells it to use the fbsource xplat toolchain which will be consistent across everybody's devserver.

Reviewed By: diliop, capickett

Differential Revision: D73453607

fbshipit-source-id: 5e2e36c7a0963642a6b565b4601a08e0bc062b45