Emit Interpreter Errors
Summary:
When evaluating a SimpliHack expression, we may encounter statements or expressions that are not supported. Currently we will silently fail to produce a prompt with no way of the user knowing what is wrong. This makes authoring new promptlets cumbersome and error prone.

This diff fixes this by having the interpreter produce a Typing_error when it gets stuck. This builds off of `EvalError` that was added in D73405694 by tracking/providing additional contextual information that is included as part of the the error.
1. We maintain a call stack that we push onto whenever a call expression is evaluated. This is used to report a stack trace as part of the error message.
2. We utilize the position information (which was previously discarded in most places), so we can refer to the particular part of the AST we got stuck on.
3. I got rid of the `Value.str` assert methods and added the helper `Expr.eval_str` instead. This is again because we want the positional information. In almost all cases the pattern was to call `Value.str @@ Expr.eval ...`, which will now be `Expr.eval_str ...`

We report these evaluation errors as part of the simplihack tast check.

Differential Revision: D73413228

fbshipit-source-id: 6453efad3de9b0003508b96cead78bb427a0adfc