Raise Errors During Evaluation
Summary:
Currently the interpreter handles errors by using `Option.Let_syntax` when encountering unexpected or unsupported expressions. We want to be able to report errors instead of silently failing. Since there isn't any mechanism to recover when encountering these unexpected results, I am switching to raising an `EvalError` exception instead.

This exception is caught at the entry point and converted to the option, so the API remains consistent. This is an intermediate step towards reporting more detailed evaluation errors, where we will store additional information in the `EvalError` in order to produce an error. I pulled this out as a separate diff first to make the next diff on the stack easier to review

Differential Revision: D73405694

fbshipit-source-id: 7b157f16636f5bfb3165c49aa3a5efd5da93fb72