Generate association from Thrift types to their schema bundles
Summary:
Introduces a new trait that contains a pointer to the schema bundle associated with a given type as well as the hashed identifiers needed to access the definition node corresponding to the generated definition.
Because this needs to work when only the _types.h header is included we can't use the existing traits for structs/enums (those are defined in the _data.h header). Once we're introducing a new trait we may as well make it homogeneous for different definition kinds

Reviewed By: praihan

Differential Revision: D72676523

fbshipit-source-id: 86c9848ae9a063fc52f668794cbc1c18c2d10254