Make `folly::get_mutable_exception` separate from `folly::get_exception`
Summary:
`get_mutable_exception` is a separate verb because:

  - Mutable exception access is rare.  It may run into thread-safety bugs if a `std::current_exception()` pointer is accessed outside of the thread that threw it -- the standard permits reference semantics here!

  - Making mutable access explicit enables no-allocations, no-atomics optimizations for the `const`-access path.

I specifically ran into the *latter* issue, since I was implementing immortal exception pointers, and there's no 0-overhead way for those to provide mutable access.

On the other hand, with `get_mutable_exception`, it's completely fine for the accessor to modify the state of the potentially-immortal exception pointer to contain a just-copied `std::exception_ptr`.

Reviewed By: ispeters

Differential Revision: D73390138

fbshipit-source-id: ebfa60d54462cd372340b76b4dad08b24096e81a