Add size budget
Summary:
We encountered several situations where the populator tries to create an unreasonably large data structure, especially for complex and nested thrift types (e.g. `facebook::tupperware::federation::FedUpdateRequest`). This resulted in either timeouts or OOMs.

This diff adds a populator option that specifies the target size limit for the populated data structure. The size limit does not account for container overhead and union tags, so the results may be slightly larger than the specified limit.

Reviewed By: Mizuchi

Differential Revision: D73199010

fbshipit-source-id: 4c2ff359e54a84136fb4dae110ba28fd0492296d