Fix crashes in gsymutil resource accounting
Summary:
poll uses wnohang option, so ocasionally we got empty stats and the service crashed because we never actually waited for the process to complete and there is a safeguard in folly for that https://fburl.com/code/ujrzktm5 We mostly succeeded probably because communicate call above returned shortly before the process death, but there were ocasional races.

We could poll in a cycle in the service until the state is running, but that seemed like a cludge. So I decided to make changes to folly.

I thought about adding options field to poll. But semantically seemed wrong that we can wait in a function named poll. Also there is no retry loop for properly handling waiting erros.

Wait already uses waitpid under the hood, changing it to wait4 seemed risky.

So I decided to add a new function which is kinda of a mix of a two.

Reviewed By: jwiepert, hodgesds

Differential Revision: D73506301

fbshipit-source-id: 344833608409a5bdd5873e0fe50b863b4347c1a5