remove now-redundant `originalFilename` field from HHBBC's function representations
Summary:
The duplication of `originalFilename` and `originalUnit` were introduced in D6770083 to support the flattening of traits in HHBBC.

A later diff, D6770160, removed the trait flattening code from hphpc, but we never appear to have cleaned up the originalFilename field.

This change makes it so we only keep `originalUnit` around, and renames `FuncEmitter`'s originalFilename to originalUnit for consistency.

As the asserts demonstrate, this saves 8 bytes from the `php::Func` structs in both low-pointer mode and regular-sized pointer modes in HHBBC. The `originalUnit` field had to be aligned above Attr where we were wasting some padding to realize the gain in lowptr mode.

Reviewed By: ricklavoie

Differential Revision: D73373868

fbshipit-source-id: 78c630124ed53b55c29b58bd49f5fdcbaace9990