let multiple automatic arenas share a common mapper callback
Summary:
We used to have only one automatic arena, but that proves insufficient on very wide machines.
We had some support for using hugetlb pages in arena 0. This diff tries to expand the support to cover all automatic arenas.
The implementation is to use extend hooks backed by a shared page mapper. The mapper supports all page sizes, and is atomic when allocating, making it OK to share the same mapper among multiple arenas.

Reviewed By: ottoni

Differential Revision: D73421329

fbshipit-source-id: 56059026bdfe34e79dc8974e879bb8c51509ff8f