fix union Type name mangling for underscore names
Summary:
The `Union.Type` api breaks in inplace for `union`s named with a leading `_`. To mimic thrift-py3, we have to generate `___UnionType` enum with 3 underscores. The assignment `_Union.Type = ___UnionType` then breaks due to python class mangling, and the current handling assumes no leading underscore.

Switch to using `_fbthrift__` prefix for referring to the `Type` enum so we're no longer sensitive to leading underscores.

Reviewed By: prakashgayasen

Differential Revision: D73524617

fbshipit-source-id: 0b42b24290f81d316e88d0170312b58d8f5bda22