Delete to_exception_ptr() && overload
Summary:
It doesn't make sense to copy the `std::exception_ptr` when `*this` is an rvalue.

The idea here is -- first, delete the `&&` overload to make sure nothing is using it. Later, if needed, it could be reintroduced as a way of unwrapping the `exception_ptr`.

#buildmore - Be thorough
ig-no-test - No need to run instagram tests; build phase is sufficient.
#buildsonlynotests - This diff has no runtime effects.

Reviewed By: ispeters

Differential Revision: D72877186

fbshipit-source-id: 8efdf463a20c0051cc47baa85792a0190f2d5563