Reorder header so types are complete when used
Summary:
The type `TypeRef` was declared early in the header, but it was later used in
a `std::vector` as part of `FunctionSink`. The type size must be known to be in
a `std::vector`, and it was not working with an incomplete type.

Reorder the `TypeRef` above the `FunctionSink` in order to have a complete type
and a size to create those `std::vector`.

Additionally, add an include to `<cstddef>` in `Math.h` because it uses the type
`ptrdiff_t` which is defined in that header.

Reviewed By: narissiam

Differential Revision: D73470750

fbshipit-source-id: f8ca8f90ea7eee1e40c5c1432529ace3fd87cf48