Implement pretty debug printing for SyntaxGraph
Summary:
Debugging `SyntaxGraph` is pretty hard today so with the recent addition of `tree_printer` to core Thrift, we can do much better!

Pretty printing for all nodes are supported with the exception of `protocol.Value` (we need to figure out a better representation first). This affects custom defaults in `FieldNode`, `Annotation`, and `ConstantNode`.

Since the graph may have cycles, we only print a `DefinitionNode` or `ProgramNode` once.

Reviewed By: Mizuchi

Differential Revision: D73387342

fbshipit-source-id: 337bd890bd044e5f8ae800546999b2b864736216