fix implicit inheritance of require class/this as constraints via use
Summary:
According to the HIP, `require class` (and `require this as`) trait constraints should be implicitly inherited by traits across `use` statements: if T1 requires class C and T2 uses T1, then implicitly T2 requires class C.  This is analogous to the behaviour of `require extends` and `require implements`.

The implementation of `require class` (and `require this as`) didn't honour this behaviour;  this diff fixes this.

Reviewed By: andrewjkennedy

Differential Revision: D73592241

fbshipit-source-id: 2ed3e4df85f67a33a54d3e8f929f5e0c43151e27