source_manager_backend
Summary:
This diff generalizes the `read_from_file_system` option that was added recently.

By abstracting away all IO into the `backend` interface, we allow the flexibility of alternate backend implementations.

For example, we may choose to implemented a backend which stores files in an in-memory map. This is different from using `add_virtual_file` because `add_virtual_file` eagerly indexes (calculates offsets, see `get_line_offsets`) for every file that is added. The approach here is a much superior implementation.

In the future `add_virtual_file` could also be superseded with a custom `backend` implementation but there are too many uses currently.

Reviewed By: vitaut

Differential Revision: D73410257

fbshipit-source-id: 6bcadeaac209bea923cc39a684df103a0ee6e9a9