Fix crash due to type lies in BuiltinEnum::coerce() wrt class types
Summary:
niharikad's recent changes revealed that BuiltinEnum::coerce() was
misbehaving. It claims to return `?arraykey`, but when passed a class, it would
return the input instead of the string classname, if the underlying type of
enum was not representable by a single datatype, e.g. if the base class of the
enum was a classname, which temporarily supports classes too (although not in
enums).

Fix this by returning the original value only if we did what the comment
claims, i.e. converted string to int.

Kill the JIT builtin optimization if the input is class, since it is wrong too.

Reviewed By: jtwarren

Differential Revision: D73804557

fbshipit-source-id: c57b85962a23189bd776f4e4de8bf25af1338bf8