Add new VerifyTypeTS instruction for inline type enforcement
Summary:
Add a new instruction `VerifyTypeTS` that will be used to build a "lie detector", channelling types from `hh` down to HackC in order to flag when a value's runtime type doesn't match the type declared or inferred in `hh`. The instruction takes a type structure and a value, and has similar behaviour to the existing `IsTypeStructC` instruction (used for `as`), except that it fails with a fatal (as with type enforcement) rather than an ordinary exception. Also, like `as`, it does "deep" checking, looking at fields of tuples and shapes, and checking enum values. Furthermore, the default is to produce only a warning; to get the fatal, set `ErrorHandling.ThrowExceptionOnInlineTypeError = true`. One small different from `as` is that it simply ignores types that are in erasure positions, rather than producing an error.

To test the instruction, we introduce a flag `Hack.Lang.CheckedUnsafeCast` that controls the insertion of a `VerifyTypeTS` instruction for uses of `HH\FIXME\UNSAFE_CAST<_,t>`. In essence this has the same effect as `as t`, but with the differences mentioned above. Regarding erasure, the `HH\FIXME\UNSAFE_CAST<mixed,vec<int>>` will have the same effect as `HH\FIXME\UNSAFE_CAST<mixed,vec<_>>`, whereas `as vec<int>` is rejected both by Hack and HHVM.

Reviewed By: jano

Differential Revision: D68771154

fbshipit-source-id: 0c8c8d37bdd31ff355c30ba6dce1ef5080ed8b43