Fix Deadlock due to Exceptions Thrown from Inline Getters
Summary:
The LRUCache allows for getters to be futures which will return at a later point in time, or a getter to be an inline lambda which will return a populated future.

In the current implementation, if something falls into the latter category and throws an exception before creating and returning the future, the exception will not get caught by the LRUCache and the thenTry logic does not run. This leaves an unfulfilled promise in the cache and all future attempts to access this key will result on waiting on that promise.  This doesn't appear to ever time out and the only way to resolve it is to restart watchman.

With this change, the getter moves to run inside a future itself so any exception caught will be forwarded to the following thenTry and handled there.  Since no executor is attached, this should run in the same context as the getter previously did.  I think.

Reviewed By: jdelliot, genevievehelsel

Differential Revision: D73710162

fbshipit-source-id: 2580451dc3ac00f8e2edd2814c62910566704b13