record.thrift
Summary:
This diff implements records as described in the Object Model. The specific implementation here is called `SerializableRecord` because it has been built for serde, *not* performance.

Actual operations on Thrift data should not use this class. Instead, an *optimized* record implementation should be used and only converted to/from `SerializableRecord` when persistence/transmission is needed.

NOTE: The `Any` record is missing at the moment. It is the most "controversial" here so I wanted to hold off on adding that until all discussion around it has resolved.

Reviewed By: aristidisp

Differential Revision: D70018523

fbshipit-source-id: 71c612af59472ae07f218e647d73920a5e51eb71