Added advanceToNext to ManualTimekeeper.
Summary:
This class is heavily used in testing.

We're using it in Delos's "Deterministic Tester", where we randomly permute execution orders and timeouts of different events to ensure our system functions correctly.

We can randomly permute timeouts significantly more effectively if we can advance to the next timeout, versus guessing when the next timeout may be.  This diff adds that functionalty.

Reviewed By: yfeldblum

Differential Revision: D70102206

fbshipit-source-id: 206813a79cebbf42ddf8415e69ea663e1e8a236b