Remove error for MAKE_IC_INACCESSSIBLE in Zoned code
Summary:
On the surface this error makes sense, but the way it's implemented is very non-ideal.

ATM it fails on code like this:

```
newctx MyCtx as [];

<<__Memoize(#MakeICInaccessible)>>
function memo_fn1($a, $b)[MyCtx]: void {
}
```

Which at runtime is actually `[defaults]` which means that the runtime current both: requires it to have a flavor when the default is disable, and also errors when any flavor is provided.

To fix this I'm just removing this error to allow `#MakeICInaccessible` to be used on zoned (and also newctx's). I will implement this "error" as a lint in WWW, since it's really a "strong" suggestion, as zoned code can be called from code that uses `#MakeICInaccessible` already (though it probably shouldn't be).

Reviewed By: viratyosin

Differential Revision: D73154908

fbshipit-source-id: 380f837680a4e3cb89b28e00230d5a270bee49d0